requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==2.0.7
django-celery==3.2.2
django-grappelli==2.11.1
easy-thumbnails==2.5
kombu==3.0.37
Pillow==5.2.0
pytz==2018.5
six==1.11.0
sorl-thumbnail==12.4.1
vine==1.1.4
amqp==1.4.9
anyjson==0.3.3
astroid==2.0.4
billiard==3.3.0.23
celery==3.1.26.post2
confusable-homoglyphs==3.2.0
Django==2.0.7
django-appconf==1.0.2
django-bootstrap-form==3.4
django-celery==3.2.2
django-grappelli==2.11.1
django-registration-redux==2.4
django-user-accounts==2.0.3
django-widget-tweaks==1.4.2
easy-thumbnails==2.5
isort==4.3.4
kombu==3.0.37
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pylint==2.1.1
pytz==2018.5
six==1.11.0
sorl-thumbnail==12.4.1
typed-ast==1.1.0
vine==1.1.4
wrapt==1.10.11
